Refactor Configurations
