+++ EyesGuard/Configurations/Configuration.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Xml;
+using System.Xml.Serialization;
+using static EyesGuard.App;
+
+namespace EyesGuard.Configurations
+{
+    public partial class Configuration
+    {
+
+        public Configuration()
+        {
+
+        }
+
+    }
+}
